Alarms file renaming
