+++ packages/yourloops/components/monitoring-alert/monitoring-alerts-content-configuration.tsx
@@ -39,8 +39,8 @@ import TextField from '@mui/material/TextField'
 import Button from '@mui/material/Button'
 import { type Monitoring } from '../../lib/team/models/monitoring.model'
 import { type Patient } from '../../lib/patient/models/patient.model'
-import useAlarmsContentConfiguration from './alarms-content-configuration.hook'
-import { buildBgValues, buildThresholds, onBasicDropdownSelect, PERCENTAGES } from './alarm-content-configuration.util'
+import useAlarmsContentConfiguration from './monitoring-alerts-content-configuration.hook'
+import { buildBgValues, buildThresholds, onBasicDropdownSelect, PERCENTAGES } from './monitoring-alert-content-configuration.util'
 import FormHelperText from '@mui/material/FormHelperText'
 import { useAuth } from '../../lib/auth'
 import { LoadingButton } from '@mui/lab'
