+++ packages/yourloops/components/patient-list/models/grid-row.model.ts
@@ -27,7 +27,7 @@
 
 import { type PatientListColumns } from '../enums/patient-list.enum'
 import { type Patient } from '../../../lib/patient/models/patient.model'
-import { type Alarms } from '../../../lib/patient/models/alarms.model'
+import { type Alarms } from '../../../lib/patient/models/monitoring-alerts.model'
 
 export interface GridRowModel {
   id: string
