+++ packages/yourloops/components/patient/utils.ts
@@ -26,7 +26,7 @@
  */
 
 import moment from 'moment-timezone' // TODO: Change moment-timezone lib with something else
-import { type Alarms } from '../../lib/patient/models/alarms.model'
+import { type Alarms } from '../../lib/patient/models/monitoring-alerts.model'
 import { type MedicalData } from '../../lib/data/models/medical-data.model'
 import { type Patient } from '../../lib/patient/models/patient.model'
 import { type MedicalTableValues } from './models/medical-table-values.model'
