+++ packages/yourloops/lib/team/models/i-team-member.model.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { type Alarms } from '../../patient/models/alarms.model'
+import { type Alarms } from '../../patient/models/monitoring-alerts.model'
 import { type Monitoring } from './monitoring.model'
 import { type TeamMemberRole } from './enums/team-member-role.enum'
 import { type UserInvitationStatus } from './enums/user-invitation-status.enum'
