+++ packages/yourloops/pages/team/team-details-page.tsx
@@ -39,7 +39,7 @@ import TeamInformation from '../../components/team/team-information'
 import TeamMembers from '../../components/team/team-members'
 import { commonComponentStyles } from '../../components/common'
 import { useAuth } from '../../lib/auth'
-import TeamAlarmsConfiguration from '../../components/team/team-alarms-configuration'
+import TeamAlarmsConfiguration from '../../components/team/team-monitoring-alerts-configuration'
 import { useSelectedTeamContext } from '../../lib/selected-team/selected-team.provider'
 import { useParams } from 'react-router-dom'
 import { setPageTitle } from '../../lib/utils'
