+++ packages/yourloops/test/unit/components/monitoring-alert/monitoring-alert-card.test.tsx
@@ -33,9 +33,9 @@ import { type RemoteMonitoringWidgetProps } from '../../../../components/dashboa
 import { createPatient, triggerMouseEvent } from '../../common/utils'
 import i18n from '../../../../lib/language'
 import * as authHookMock from '../../../../lib/auth'
-import AlarmCard from '../../../../components/alarm/alarm-card'
+import AlarmCard from '../../../../components/monitoring-alert/monitoring-alert-card'
 import type User from '../../../../lib/auth/models/user.model'
-import { type Alarms } from '../../../../lib/patient/models/alarms.model'
+import { type Alarms } from '../../../../lib/patient/models/monitoring-alerts.model'
 import { type Monitoring } from '../../../../lib/team/models/monitoring.model'
 import { getTheme } from '../../../../components/theme'
 import { Unit } from 'medical-domain'
