+++ packages/yourloops/test/unit/components/monitoring-alert/monitoring-alert-content-configuration.util.test.ts
@@ -28,7 +28,7 @@ import {
   buildBgValues,
   buildThresholds,
   getConvertedValue
-} from '../../../../components/alarm/alarm-content-configuration.util'
+} from '../../../../components/monitoring-alert/monitoring-alert-content-configuration.util'
 import { Unit } from 'medical-domain'
 
 describe('AlarmsContentConfiguration util', function () {
